+++ app/src/main/kotlin/com/pitchedapps/frost/parsers/FrostParser.kt
@@ -82,16 +82,16 @@ internal fun <T> List<T>.toJsonString(tag: String, indent: Int) = StringBuilder(
  */
 internal abstract class FrostParserBase<out T : Any>(private val redirectToText: Boolean) : FrostParser<T> {
 
-    override final fun parse(cookie: String?) = parseFromUrl(cookie, url)
+    final override fun parse(cookie: String?) = parseFromUrl(cookie, url)
 
-    override final fun parseFromData(cookie: String?, text: String): ParseResponse<T>? {
+    final override fun parseFromData(cookie: String?, text: String): ParseResponse<T>? {
         cookie ?: return null
         val doc = textToDoc(text) ?: return null
         val data = parseImpl(doc) ?: return null
         return ParseResponse(cookie, data)
     }
 
-    override final fun parseFromUrl(cookie: String?, url: String): ParseResponse<T>? =
+    final override fun parseFromUrl(cookie: String?, url: String): ParseResponse<T>? =
             parse(cookie, frostJsoup(cookie, url))
 
     override fun parse(cookie: String?, document: Document): ParseResponse<T>? {